chore(separate file): make test and routes into diff module
[finishes #166660872]
