--- src/data/cars.js
@@ -1,4 +1,0 @@
-// jshint esversion: 6
-const carDb = [];
-
-module.exports = carDb;
