--- src/data/flags.js
@@ -1,14 +1,0 @@
-// jshint esversion: 6
-
-const flagDb = [
-  {
-    id: 1,
-    userId: 4,
-    carId: 3,
-    reason: 'pricing',
-    description: 'lorem ipesum',
-    createdOn: '18-06-05',
-  },
-];
-
-module.exports = flagDb;
