--- src/data/orders.js
@@ -1,15 +1,0 @@
-// jshint esversion : 6
-
-const orderDb = [
-  {
-    id: 1,
-    buyer: 5,
-    carId: 6,
-    price: 100000,
-    priceOffered: 50000,
-    status: 'pending',
-    createdOn: '2018-05-06',
-  },
-];
-
-module.exports = orderDb;
