--- src/data/users.js
@@ -1,14 +1,0 @@
-// jshint esversion: 6
-const userDb = [
-  {
-    id: 1,
-    email: 'johndoe@gmail.com',
-    firstName: 'john',
-    lastName: 'doe',
-    password: 'johnpassword',
-    address: 'Congo, DRC',
-    isAdmin: true,
-  },
-];
-
-module.exports = userDb;
